Revert "init admin user in aslan (#2867)"
This reverts commit b2051dadfd001cf9a59b11c7d3e8512dc5da02f4.
