+++ pkg/cli/initconfig/cmd/init.go
@@ -18,6 +18,8 @@ package cmd
 
 import (
 	_ "embed"
+	"encoding/base64"
+	"encoding/json"
 	"time"
 
 	"github.com/spf13/cobra"
@@ -25,7 +27,10 @@ import (
 	"github.com/koderover/zadig/pkg/config"
 	"github.com/koderover/zadig/pkg/setting"
 	"github.com/koderover/zadig/pkg/shared/client/aslan"
+	"github.com/koderover/zadig/pkg/shared/client/policy"
+	"github.com/koderover/zadig/pkg/shared/client/user"
 	kubeclient "github.com/koderover/zadig/pkg/shared/kube/client"
+	"github.com/koderover/zadig/pkg/tool/httpclient"
 	"github.com/koderover/zadig/pkg/tool/kube/updater"
 	"github.com/koderover/zadig/pkg/tool/log"
 )
@@ -65,6 +70,21 @@ func run() error {
 }
 
 func initSystemConfig() error {
+	email := config.AdminEmail()
+	password := config.AdminPassword()
+	domain := config.SystemAddress()
+
+	uid, err := presetSystemAdmin(email, password, domain)
+	if err != nil {
+		log.Errorf("presetSystemAdmin err:%s", err)
+		return err
+	}
+
+	if err := presetRoleBinding(uid); err != nil {
+		log.Errorf("presetRoleBinding err:%s", err)
+		return err
+	}
+
 	if err := createLocalCluster(); err != nil {
 		log.Errorf("createLocalCluster err:%s", err)
 		return err
@@ -90,6 +110,73 @@ func scaleWarpdrive() error {
 
 	log.Errorf("Failed to get workflow concurrency settings, error: %s", err)
 	return err
+}
+
+func presetSystemAdmin(email string, password, domain string) (string, error) {
+	r, err := user.New().SearchUser(&user.SearchUserArgs{Account: setting.PresetAccount})
+	if err != nil {
+		log.Errorf("SearchUser err:%s", err)
+		return "", err
+	}
+	if len(r.Users) > 0 {
+		log.Infof("User admin exists, skip it.")
+		return r.Users[0].UID, nil
+	}
+	user, err := user.New().CreateUser(&user.CreateUserArgs{
+		Name:     setting.PresetAccount,
+		Password: password,
+		Account:  setting.PresetAccount,
+		Email:    email,
+	})
+	if err != nil {
+		log.Errorf("created  admin err:%s", err)
+		return "", err
+	}
+	// report register
+	err = reportRegister(domain, email)
+	if err != nil {
+		log.Errorf("reportRegister err: %s", err)
+	}
+	return user.Uid, nil
+}
+
+type Operation struct {
+	Data string `json:"data"`
+}
+type Register struct {
+	Domain    string `json:"domain"`
+	Username  string `json:"username"`
+	Email     string `json:"email"`
+	CreatedAt int64  `json:"created_at"`
+}
+
+func reportRegister(domain, email string) error {
+	register := Register{
+		Domain:    domain,
+		Username:  "admin",
+		Email:     email,
+		CreatedAt: time.Now().Unix(),
+	}
+	registerByte, _ := json.Marshal(register)
+	encrypt, err := RSAEncrypt([]byte(registerByte))
+	if err != nil {
+		log.Errorf("RSAEncrypt err: %s", err)
+		return err
+	}
+	encodeString := base64.StdEncoding.EncodeToString(encrypt)
+	reqBody := Operation{Data: encodeString}
+	_, err = httpclient.Post("https://api.koderover.com/api/operation/admin/user", httpclient.SetBody(reqBody))
+	return err
+}
+
+func presetRoleBinding(uid string) error {
+	return policy.NewDefault().CreateOrUpdateSystemRoleBinding(&policy.RoleBinding{
+		Name: config.RoleBindingNameFromUIDAndRole(uid, setting.SystemAdmin, "*"),
+		UID:  uid,
+		Role: string(setting.SystemAdmin),
+		Type: setting.ResourceTypeSystem,
+	})
+
 }
 
 func createLocalCluster() error {
